+++ src/east/east_yml.py
@@ -55,29 +55,52 @@ def load_east_yml(project_dir: str):
                 " in [bold]east.yml[/]!"
             )
 
-    check_duplicated_entries(east_yml["apps"], "apps", "name")
-    check_duplicated_entries(east_yml["samples"], "samples", "name")
-    for app in east_yml["apps"]:
-        check_duplicated_entries(
-            app["build-types"], f"{app['name']}.build-types", "type"
+    # This is needed to discern between apps key present, apps key present, but not set
+    # and not app key at all.
+    apps = east_yml.get("apps", -1)
+
+    # apps are optional
+    if apps is not None and apps != -1:
+        check_duplicated_entries(east_yml["apps"], "apps", "name")
+        for app in east_yml["apps"]:
+            check_duplicated_entries(
+                app["build-types"], f"{app['name']}.build-types", "type"
+            )
+
+    if apps is None:
+        # Exists, but it was not set
+        raise EastYmlLoadError(
+            "[bold]apps[/] key in [bold yellow]east.yml[/] has no apps listed under it!"
         )
 
-    # For each sample that has inherit key check if that app with that build type exists
-    for sample in east_yml["samples"]:
-        if "inherit-build-type" in sample:
-            inherited_app = sample["inherit-build-type"]["app"]
-            inherited_type = sample["inherit-build-type"]["build-type"]
+    # samples are optional
+    if east_yml.get("samples"):
+        check_duplicated_entries(east_yml["samples"], "samples", "name")
+        # For each sample that has inherit key check if that app with that build type
+        # exists
+        for sample in east_yml["samples"]:
+            if "inherit-build-type" in sample:
+                # Inherit needs apps
+                if not east_yml.get("apps"):
+                    raise EastYmlLoadError(
+                        f"Sample [bold]{sample['name']}[/] is trying to inherit, but"
+                        " there are no apps to inherit from!"
+                    )
 
-            app = return_dict_on_match(east_yml["apps"], "name", inherited_app)
-            if not app:
-                raise EastYmlLoadError(
-                    f"Sample [bold]{sample['name']}[/] is trying to inherit from a"
-                    f" [bold red]non-existing[/] app [bold]{inherited_app}[/]."
-                )
-            if not return_dict_on_match(app["build-types"], "type", inherited_type):
-                raise EastYmlLoadError(
-                    f"Sample [bold]{sample['name']}[/] is trying to inherit from a"
-                    f" [bold red]non-existing[/] build-type [bold]{inherited_type}[/]."
-                )
+                inherited_app = sample["inherit-build-type"]["app"]
+                inherited_type = sample["inherit-build-type"]["build-type"]
+
+                app = return_dict_on_match(east_yml["apps"], "name", inherited_app)
+                if not app:
+                    raise EastYmlLoadError(
+                        f"Sample [bold]{sample['name']}[/] is trying to inherit from a"
+                        f" [bold red]non-existing[/] app [bold]{inherited_app}[/]."
+                    )
+                if not return_dict_on_match(app["build-types"], "type", inherited_type):
+                    raise EastYmlLoadError(
+                        f"Sample [bold]{sample['name']}[/] is trying to inherit from a"
+                        " [bold red]non-existing[/] build-type"
+                        f" [bold]{inherited_type}[/]."
+                    )
 
     return east_yml