+++ CHANGELOG.md
@@ -8,8 +8,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Changed
 
-- Make east.yml optional for everything, expect for the usage of east build
+- Make east.yml optional for everything, except for the usage of east build
   command with --build-type option.
+- Make apps key and samples key inside east.yml optional. This is useful for
+  driver projects, which do not need apps, or any project that might not have
+  samples.
+
+### Fixed
+
+- Properly handle _east build_ commands outside of applications and samples.
+  This means that running _east build_ command will default to plain west
+  behaviour, as it should.
 
 ## [0.4.0] - 2022-11-21